+++ packages/example-nextjs/src/app/Form-for-test.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React from 'react';
-import { useUrlState } from 'state-in-url';
+import { useUrlState } from 'state-in-url/next';
 
 import { Field } from './components/Field';
 import { Input } from './components/Input';
