+++ packages/example-nextjs/src/app/Status-for-test.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React from 'react';
-import { useUrlState } from 'state-in-url';
+import { useUrlState } from 'state-in-url/next';
 
 import { form } from './form';
