+++ packages/example-nextjs/src/app/Status.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React from 'react';
-import { useUrlState } from 'state-in-url';
+import { useUrlState } from 'state-in-url/next';
 
 import { SourceCodeBtn } from './components/SourceCodeBtn';
 import { form } from './form';
