+++ packages/urlstate/index.ts
@@ -1,8 +1,8 @@
 import { decode, encode, parseSsrQs } from './encoder';
 import { decodeState, encodeState } from './encodeState';
+import { useUrlState } from './next/useUrlState';
 import { useSharedState } from './useSharedState';
 import { useUrlEncode } from './useUrlEncode';
-import { useUrlState } from './useUrlState';
 import { useUrlStateBase } from './useUrlStateBase';
 import { isSSR, type JSONCompatible, type Type, typeOf } from './utils';
 export {
