+++ packages/urlstate/next/useUrlState/useUrlState.ts
@@ -1,8 +1,10 @@
 import { useRouter } from 'next/navigation';
+import { useSearchParams } from 'next/navigation';
 import React from 'react';
 
-import { useUrlStateBase } from '../useUrlStateBase';
-import { type DeepReadonly, type JSONCompatible } from '../utils';
+import { parseSsrQs } from '../../encoder';
+import { useUrlStateBase } from '../../useUrlStateBase';
+import { type DeepReadonly, type JSONCompatible } from '../../utils';
 
 /**
  * NextJS hook. Returns `state`, `updateState`, and `updateUrl` functions
@@ -14,13 +16,15 @@ import { type DeepReadonly, type JSONCompatible } from '../utils';
  * ```ts
  * export const form = { name: '' };
  * const { state, updateState, updateUrl } = useUrlState(form);
+ * // for nextjs seerver components
+ * // const { state, updateState, updateUrl } = useUrlState(form, searchParams);
  *
  * updateState({ name: 'test' });
  * // by default it's uses router.push with scroll: false
  * updateUrl({ name: 'test' }, { replace: true, scroll: true });
  *  ```
  *
- *  * Github {@link https://github.com/asmyshlyaev177/state-in-url/tree/main/packages/urlstate/useUrlState#api}
+ *  * Github {@link https://github.com/asmyshlyaev177/state-in-url/tree/main/packages/urlstate/next/useUrlState#api}
  */
 export function useUrlState<T extends JSONCompatible>(
   defaultState: T,
@@ -44,6 +48,14 @@ export function useUrlState<T extends JSONCompatible>(
     [updateUrlBase],
   );
 
+  const sp = useSearchParams();
+
+  React.useEffect(() => {
+    updateState(
+      parseSsrQs(Object.fromEntries([...sp.entries()]), defaultState),
+    );
+  }, [sp]);
+
   return {
     /**
      * * Example:
@@ -53,7 +65,7 @@ export function useUrlState<T extends JSONCompatible>(
      * updateState(curr => ({ ...curr, name: 'test' }) );
      *  ```
      *
-     *  * Github {@link https://github.com/asmyshlyaev177/state-in-url/tree/main/packages/urlstate/useUrlState#updatestate}
+     *  * Github {@link https://github.com/asmyshlyaev177/state-in-url/tree/main/packages/urlstate/next/useUrlState#updatestate}
      */
     updateState,
     /**
@@ -64,7 +76,7 @@ export function useUrlState<T extends JSONCompatible>(
      * updateUrl(curr => ({ ...curr, name: 'test' }), { replace: true, scroll: false  } );
      *  ```
      *
-     *  * Github {@link https://github.com/asmyshlyaev177/state-in-url/tree/main/packages/urlstate/useUrlState#updateurl}
+     *  * Github {@link https://github.com/asmyshlyaev177/state-in-url/tree/main/packages/urlstate/next/useUrlState#updateurl}
      */
     updateUrl,
     state: state as DeepReadonly<typeof state>,
